Home

WorldMapData.jl Documentation

Welcome to the WorldMapData.jl documentation! This package provides access to GSHHS world map data. The data provides information on landmass, water, and political boundaries.

You can find and download the original dataset from here https://www.soest.hawaii.edu/pwessel/gshhg/

The package provides the following function to interact with the dataset:

WorldMapData.valid_resolution — Function.

Return whether the input symbol is a valid resolution for GSHHS data.

Arguments: - res::Symbol Symbol to validate

Returns: - valid::Bool Boolean of resolution validity.

source

WorldMapData.load_GSHHS_data — Function.

Load GSHHS Shapefile data and return all polygons in the data set.

All data sets come in 5 different resolutions: - :FULL: Full resolution. These contain the maximum resolution of this data and has not been decimated. - :HIGH: High resolution. The Douglas-Peucker line reduction was used to reduce data size by ~80% relative to full. - :INTERMEDIATE: Intermediate resolution. The Douglas-Peucker line reduction was used to reduce data size by ~80% relative to high. - :LOW: Low resolution. The Douglas-Peucker line reduction was used to reduce data size by ~80% relative to intermediate. - :CRUDE: Crude resolution. The Douglas-Peucker line reduction was used to reduce data size by ~80% relative to low.

For each resolution there are several levels; these depends on the data type.

The shoreline data are distributed in 6 levels: - 1: Continental land masses and ocean islands, except Antarctica. - 2: Lakes - 3: Islands in lakes - 4: Ponds in islands within lakes - 5: Antarctica based on ice front boundary. - 6: Antarctica based on grounding line boundary.

Note that because GIS software confusingly seem to assume a Cartesian geometry, any polygon straddling the Dateline is broken into an east and west component. The most obvious example is Antarctica.

source

WorldMapData.load_WDBII_border_data — Function.

Load WDBI Shapefile data and return all polygons in the data set.

All data sets come in 5 different resolutions: - :FULL: Full resolution. These contain the maximum resolution of this data and has not been decimated. - :HIGH: High resolution. The Douglas-Peucker line reduction was used to reduce data size by ~80% relative to full. - :INTERMEDIATE: Intermediate resolution. The Douglas-Peucker line reduction was used to reduce data size by ~80% relative to high. - :LOW: Low resolution. The Douglas-Peucker line reduction was used to reduce data size by ~80% relative to intermediate. - :CRUDE: Crude resolution. The Douglas-Peucker line reduction was used to reduce data size by ~80% relative to low.

For each resolution there are several levels; these depends on the data type.

The political boundary data come in 3 levels: - 1: National boundaries. - 2: Internal (state) boundaries for the 8 largest countries only. - 3: Maritime boundaries.

source

WorldMapData.load_WDBII_river_data — Function.

Load WDBI Shapefile data and return all polygons in the data set.

All data sets come in 5 different resolutions: - :FULL: Full resolution. These contain the maximum resolution of this data and has not been decimated. - :HIGH: High resolution. The Douglas-Peucker line reduction was used to reduce data size by ~80% relative to full. - :INTERMEDIATE: Intermediate resolution. The Douglas-Peucker line reduction was used to reduce data size by ~80% relative to high. - :LOW: Low resolution. The Douglas-Peucker line reduction was used to reduce data size by ~80% relative to intermediate. - :CRUDE: Crude resolution. The Douglas-Peucker line reduction was used to reduce data size by ~80% relative to low.

For each resolution there are several levels; these depends on the data type.

The river database come with 11 levels: - 1: Double-lined rivers (river-lakes). - 2: Permanent major rivers. - 3: Additional major rivers. - 4: Additional rivers. - 5: Minor rivers. - 6: Intermittent rivers - major. - 7: Intermittent rivers - additional. - 8: Intermittent rivers - minor. - 9: Major canals. - 10: Minor canals. - 11: Irrigation canals.

source